requirements.txt
matplotlib==3.3.3
numpy==1.19.2
Pillow==8.1.0
pip==20.3.3
torch==1.7.1
torchvision==0.8.2
tqdm==4.56.0
urllib3==1.26.3
wheel==0.36.2

